requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
anyio>=4.8.0,<5.0.0
httpx>=0.28.1,<1.0.0
groq==0.11.0
requests==2.32.3
PyGithub==2.4.0
python-dotenv==1.0.1
tenacity==9.0.0